+++ src/api/census.ts
@@ -62,9 +62,9 @@ export class CensusOffChainApi {
      * @param managerPublicKeys ECDSA public key(s) that can manage this census
      * @param gateway A Gateway instance connected to a remote Gateway
      * @param walletOrSigner
-     * @returns Promise resolving with the new merkleRoot
+     * @returns Promise resolving with the new censusRoot
      */
-    static async addCensus(censusName: string, managerPublicKeys: string[], walletOrSigner: Wallet | Signer, gateway: IGateway | IGatewayPool): Promise<{ censusId: string, merkleRoot: string }> {
+    static async addCensus(censusName: string, managerPublicKeys: string[], walletOrSigner: Wallet | Signer, gateway: IGateway | IGatewayPool): Promise<{ censusId: string, censusRoot: string }> {
         if (!censusName || !managerPublicKeys || !managerPublicKeys.length || !gateway) return Promise.reject(new Error("Invalid parameters"))
         else if (!(gateway instanceof Gateway || gateway instanceof GatewayPool)) return Promise.reject(new Error("Invalid Gateway object"))
         else if (!walletOrSigner || !walletOrSigner._isSigner) return Promise.reject(new Error("Invalid WalletOrSinger object"))
@@ -77,7 +77,7 @@ export class CensusOffChainApi {
             // TODO: normalize the `censusId` parameter value
             // Pass the full censusId instead of the second term only
             existingRoot = await CensusOffChainApi.getRoot(censusId, gateway)
-            if (typeof existingRoot == "string" && existingRoot.match(/^0x[0-9a-zA-Z]+$/)) return { censusId, merkleRoot: existingRoot }
+            if (typeof existingRoot == "string" && existingRoot.match(/^0x[0-9a-zA-Z]+$/)) return { censusId, censusRoot: existingRoot }
         } catch (error) {
             // If it errors because it doesn't exist, we continue below
         }
@@ -85,8 +85,8 @@ export class CensusOffChainApi {
         try {
             const censusIdSuffix = CensusOffChainApi.generateCensusIdSuffix(censusName)
             const response = await gateway.sendRequest({ method: "addCensus", censusId: censusIdSuffix, pubKeys: managerPublicKeys }, walletOrSigner)
-            const merkleRoot = await CensusOffChainApi.getRoot(response.censusId, gateway)
-            return { censusId: response.censusId, merkleRoot }
+            const censusRoot = await CensusOffChainApi.getRoot(response.censusId, gateway)
+            return { censusId: response.censusId, censusRoot }
         } catch (error) {
             const message = (error.message) ? "The census could not be created: " + error.message : "The census could not be created "
             throw new Error(message)
@@ -102,7 +102,7 @@ export class CensusOffChainApi {
      * @param digested Set to true if the claim is already hashed. False otherwise.
      * @param gateway A Gateway instance pointing to a remote Gateway
      * @param walletOrSigner
-     * @returns Promise resolving with the new merkleRoot
+     * @returns Promise resolving with the new censusRoot
      */
     static addClaim(censusId: string, claim: { key: string, value?: string }, digested: boolean, walletOrSigner: Wallet | Signer, gateway: IGateway | IGatewayPool): Promise<string> {
         if (!censusId || !claim || !claim.key || !claim.key.length || !gateway) return Promise.reject(new Error("Invalid parameters"))
@@ -129,9 +129,9 @@ export class CensusOffChainApi {
      * @param digested Set to true if the claims are already hashed. False otherwise.
      * @param gateway A Gateway instance pointing to a remote Gateway
      * @param walletOrSigner
-     * @returns Promise resolving with the new merkleRoot
+     * @returns Promise resolving with the new censusRoot
      */
-    static async addClaimBulk(censusId: string, claimList: { key: string, value?: string }[], digested: boolean, walletOrSigner: Wallet | Signer, gateway: IGateway | IGatewayPool): Promise<{ merkleRoot: string, invalidClaims: any[] }> {
+    static async addClaimBulk(censusId: string, claimList: { key: string, value?: string }[], digested: boolean, walletOrSigner: Wallet | Signer, gateway: IGateway | IGatewayPool): Promise<{ censusRoot: string, invalidClaims: any[] }> {
         if (!censusId || !claimList || !claimList.length || !gateway) return Promise.reject(new Error("Invalid parameters"))
         else if (!(gateway instanceof Gateway || gateway instanceof GatewayPool)) return Promise.reject(new Error("Invalid Gateway object"))
         else if (!walletOrSigner || !walletOrSigner._isSigner) return Promise.reject(new Error("Invalid WalletOrSinger object"))
@@ -145,9 +145,9 @@ export class CensusOffChainApi {
             invalidClaims = invalidClaims.concat(partialInvalidClaims)
         }
 
-        const merkleRoot = await CensusOffChainApi.getRoot(censusId, gateway)
+        const censusRoot = await CensusOffChainApi.getRoot(censusId, gateway)
 
-        return { merkleRoot, invalidClaims }
+        return { censusRoot, invalidClaims }
     }
 
     private static addClaimChunk(censusId: string, claimList: { key: string, value?: string }[], digested: boolean, walletOrSigner: Wallet | Signer, gateway: IGateway | IGatewayPool): Promise<any[]> {
@@ -174,7 +174,7 @@ export class CensusOffChainApi {
      * Asks the Gateway to fetch
      * @param censusId The full Census ID to fetch from the Gateway
      * @param gateway A Gateway instance pointing to a remote Gateway
-     * @returns Promise resolving with the merkleRoot
+     * @returns Promise resolving with the censusRoot
      */
     static getRoot(censusId: string, gateway: IGateway | IGatewayPool): Promise<string> {
         if (!censusId || !gateway) return Promise.reject(new Error("Invalid parameters"))
@@ -182,7 +182,7 @@ export class CensusOffChainApi {
 
         return gateway.sendRequest({ method: "getRoot", censusId })
             .then(response => {
-                if (!response.root) throw new Error("The census merkle root could not be fetched")
+                if (!response.root) throw new Error("The census census root could not be fetched")
                 return response.root
             }).catch(error => {
                 const message = (error.message) ? error.message : "The request could not be completed"
@@ -191,7 +191,7 @@ export class CensusOffChainApi {
     }
 
     /**
-     * Get the number of people in the census with the given Merkle Root Hash
+     * Get the number of people in the census with the given census Merkle Root Hash
      * @param censusRootHash The Merkle Root of the census to fetch from the Gateway
      * @param dvoteGw A Gateway instance pointing to a remote Gateway
      * @returns Promise resolving with the census size
@@ -265,7 +265,7 @@ export class CensusOffChainApi {
             })
     }
 
-    /** Only works with specific merkletree format used by dump method. To add a list of plain claims use `addClaimBulk` instead */
+    /** Only works with specific merkle tree format used by dump method. To add a list of plain claims use `addClaimBulk` instead */
     static importDump() {
         // TODO: Not implemented
         throw new Error("TODO: Unimplemented")
@@ -295,7 +295,7 @@ export class CensusOffChainApi {
     }
 
     /**
-     * Fetch the proof of the given claim on the given census merkleTree using the given gateway
+     * Fetch the proof of the given claim on the given census merkle Tree using the given gateway
      * @param censusRoot The Merkle Root of the Census to query
      * @param base64Claim Base64-encoded claim of the leaf to request
      * @param gateway
@@ -310,7 +310,7 @@ export class CensusOffChainApi {
             digested: isDigested,
             claimData: base64Claim,
         }).then(response => {
-            if (typeof response.siblings != "string" || !response.siblings.length) throw new Error("The Merkle Proof could not be fetched")
+            if (typeof response.siblings != "string" || !response.siblings.length) throw new Error("The census proof could not be fetched")
             return response.siblings
         }).catch((error) => {
             const message = (error.message) ? error.message : "The request could not be completed"